# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reels App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reels:
    get:
      operationId: listReels
      tags: [reels]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReelList"

    post:
      operationId: createReel
      tags: [reels]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
  /reels/{id}:
    get:
      operationId: getReel
      tags: [reels]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"

    delete:
      operationId: deleteReel
      tags: [reels]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reels/{id}/like:
    post:
      operationId: likeReel
      tags: [reels]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Liked

    delete:
      operationId: unlikeReel
      tags: [reels]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unliked
  /reels/{id}/view:
    post:
      operationId: viewReel
      tags: [reels]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Recorded

  # ── Posts ────────────────────────────────────────────────────────────────────
...
  /store/cards:
    get:
      operationId: listGiftCards
      tags: [store]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCard"

    post:
      operationId: createGiftCard
      tags: [store]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"
  /store/cards/{id}:
    patch:
      operationId: updateGiftCard
      tags: [store]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCardUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"

    delete:
      operationId: deleteGiftCard
      tags: [store]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /store/purchase:
    post:
      operationId: purchaseGiftCard
      tags: [store]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400":
          description: Insufficient points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /store/purchases:
    get:
      operationId: listPurchases
      tags: [store]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
  /store/balance:
    get:
      operationId: getPointsBalance
      tags: [store]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"
  /store/award:
    post:
      operationId: awardPoints
      tags: [store]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsAward"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"

  # ── Leaderboard ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    UserPublic:
      type: object
      required: [id, username, displayName]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
...
    Reel:
      type: object
      required: [id, userId, views, likesCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/UserPublic"
        description:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        views:
          type: integer
        likesCount:
          type: integer
        music:
          type: ["string", "null"]
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ReelInput:
      type: object
      properties:
        description:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        music:
          type: ["string", "null"]
    ReelList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Reel"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Posts ──────────────────────────────────────────────────────────────────
...
    GiftCard:
      type: object
      required: [id, brand, category, value, pointsCost, minLevel, gradientFrom, gradientTo, emoji, description, isActive]
      properties:
        id:
          type: integer
        brand:
          type: string
        category:
          type: string
        value:
          type: string
        pointsCost:
          type: integer
        minLevel:
          type: integer
        gradientFrom:
          type: string
        gradientTo:
          type: string
        emoji:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    GiftCardInput:
      type: object
      required: [brand, category, value, pointsCost, minLevel, gradientFrom, gradientTo, emoji, description]
      properties:
        brand:
          type: string
        category:
          type: string
        value:
          type: string
        pointsCost:
          type: integer
        minLevel:
          type: integer
        gradientFrom:
          type: string
        gradientTo:
          type: string
        emoji:
          type: string
        description:
          type: string
    GiftCardUpdate:
      type: object
      properties:
        brand:
          type: string
        category:
          type: string
        value:
          type: string
        pointsCost:
          type: integer
        minLevel:
          type: integer
        gradientFrom:
          type: string
        gradientTo:
          type: string
        emoji:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    Purchase:
      type: object
      required: [id, userId, giftCardId, code, pointsSpent, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        giftCardId:
          type: integer
        giftCard:
          $ref: "#/components/schemas/GiftCard"
        code:
          type: string
        pointsSpent:
          type: integer
        createdAt:
          type: string
          format: date-time
    PurchaseInput:
      type: object
      required: [giftCardId]
      properties:
        giftCardId:
          type: integer
    PointsBalance:
      type: object
      required: [userId, balance, level]
      properties:
        userId:
          type: integer
        balance:
          type: integer
        level:
          type: integer
    PointsAward:
      type: object
      required: [userId, amount, reason]
      properties:
        userId:
          type: integer
        amount:
          type: integer
        reason:
          type: string
...