# lib/api-zod/src/generated/types/giftCardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Reels App API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCardInput {
  brand: string;
  category: string;
  value: string;
  pointsCost: number;
  minLevel: number;
  gradientFrom: string;
  gradientTo: string;
  emoji: string;
  description: string;
}