# lib/api-zod/src/generated/types/purchase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Reels App API
 * OpenAPI spec version: 0.1.0
 */
import type { GiftCard } from './giftCard';

export interface Purchase {
  id: number;
  userId: number;
  giftCardId: number;
  giftCard?: GiftCard;
  code: string;
  pointsSpent: number;
  createdAt: Date;
}